MAIN-911 RSpec recommends not using global DSL for its keywords
Historically, RSpec added key words to the global Class definition.
As of V3, they recommend not doing that, because arbitrarily monkey
patching things can lead to confusion, and unpredictable results.

Any projects that force this setting will not be able use previous
Nitra versions. Ensure we are forward and backward compatible.

For more information see:
https://relishapp.com/rspec/rspec-core/v/3-0/docs/configuration/global-namespace-dsl